+++ src/shared/preferences/language/ui/LogInAndLogOutMenuItem.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { DropdownMenuItem } from '@/shared/ui/molecules/dropdown-menu'
-import { USER_IS_IN } from '../../shared/ui/organisms/userIsIn'
+import { USER_IS_IN } from '@/shared/ui/organisms/userIsIn'
 import { TransitionLink } from '@/shared/ui/atoms/TransitionLink'
 import { LogInIcon, LogOutIcon } from '@/shared/ui/atoms/icons'
 import { useTranslation } from 'react-i18next'
